Finished Part4/4.15-4.22
Need to write better comments for edge cases , rewrite tests to reflect new code and finish 4.23

+++ part4/controllers/blogs.js
@@ -3,13 +3,6 @@ const Blog = require("../models/blog")
 const jwt = require("jsonwebtoken")
 const User = require("../models/user")
 
-const getTokenFrom = request => {
-  const authorization = request.get("authorization")
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "")
-  }
-  return null
-}
 
 
 blogRouter.get("/", async (request, response) => {
@@ -17,7 +10,19 @@ blogRouter.get("/", async (request, response) => {
   response.json(blogs)
 })
 
-blogRouter.delete("/:id", (request, response) => {
+blogRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+  const user = request.user
+  if(!user) {
+    return response.status(404).json({ error: "User not found" }).end()
+  }
+  const userinfo = await Blog.findById({ _id: request.params.id })
+  if (userinfo.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: "Unauthorized" })
+  }
   Blog
     .findByIdAndRemove(request.params.id)
     .then(() => {
@@ -43,43 +48,64 @@ blogRouter.post("/", async (request, response) => {
   const body = request.body
 
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token invalid" })
-  }
-  const user = await User.findById(decodedToken.id)
-  console.log(user)
-  if (!user) {
-    return response.status(404).json({ error: "User not found" }).end()
-  }
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user:user.id
-  })
-  if(!blog.likes) {
-    blog.likes = 0
-  }
-  if(!blog.title || !blog.url || !blog.author) {
-    return response.status(400).json({ error: "Bad Request" }).end()
+  } else {
+    const user = request.user
+    if (!user) {
+      return response.status(404).json({ error: "User not found" }).end()
+    }
 
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user.id
+    })
+
+    if (!blog.likes) {
+      blog.likes = 0
+    }
+
+    if (!blog.title || !blog.url || !blog.author) {
+      return response.status(400).json({ error: "Bad Request" }).end()
+    }
+
+    const savedBlog = await blog.save()
+    user.Blog = user.Blog.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog.toJSON())
   }
-  const savedBlog = await blog.save()
-  user.Blog = user.Blog.concat(savedBlog._id)
-  await user.save()
-  response.status(201).json(savedBlog.toJSON())
-})
-  // const user = await User.findById(request.body.userId)
-  // blog.user = user._id
-  // blog
-  //   .save()
-  //   .then(result => {
-  //     response.status(201).json(result)
-  //   })
+
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: "token invalid" })
+  // }
+  // const user = await User.findById(decodedToken.id)
+
+  // if (!user) {
+  //   return response.status(404).json({ error: "User not found" }).end()
+  // }
+  // const blog = new Blog({
+  //   title: body.title,
+  //   author: body.author,
+  //   url: body.url,
+  //   likes: body.likes,
+  //   user:user.id
+  // })
+  // if(!blog.likes) {
+  //   blog.likes = 0
+  // }
+  // if(!blog.title || !blog.url || !blog.author) {
+  //   return response.status(400).json({ error: "Bad Request" }).end()
+
+  // }
   // const savedBlog = await blog.save()
-  // user.blogs = user.blogs.concat(savedBlog._id)
+  // user.Blog = user.Blog.concat(savedBlog._id)
   // await user.save()
+  // response.status(201).json(savedBlog.toJSON())
+})
+
 
 module.exports = blogRouter
